+++ CHANGELOG.md
@@ -2,6 +2,111 @@
 
 All notable changes to this project will be documented in this file.
 
+### [4.23.3](https://github.com/wandb/terraform-aws-wandb/compare/v4.23.2...v4.23.3) (2024-09-12)
+
+
+### Bug Fixes
+
+* Allow the admin arn to be set for the KMS key ([#277](https://github.com/wandb/terraform-aws-wandb/issues/277)) ([ce69cdd](https://github.com/wandb/terraform-aws-wandb/commit/ce69cdd1d5f3f9200dcf474670910406bcb9394d))
+
+### [4.23.2](https://github.com/wandb/terraform-aws-wandb/compare/v4.23.1...v4.23.2) (2024-09-12)
+
+
+### Bug Fixes
+
+* Bump operator chart and controller image ([#276](https://github.com/wandb/terraform-aws-wandb/issues/276)) ([f010519](https://github.com/wandb/terraform-aws-wandb/commit/f010519cb1ade5c71e002a860f74c12d4d09d474))
+
+### [4.23.1](https://github.com/wandb/terraform-aws-wandb/compare/v4.23.0...v4.23.1) (2024-08-29)
+
+
+### Bug Fixes
+
+* Cleanup unneeded otel configs ([#259](https://github.com/wandb/terraform-aws-wandb/issues/259)) ([67f6fbe](https://github.com/wandb/terraform-aws-wandb/commit/67f6fbeb310da3e33d726d625bda8ad6b7baf8b1))
+
+## [4.23.0](https://github.com/wandb/terraform-aws-wandb/compare/v4.22.3...v4.23.0) (2024-08-23)
+
+
+### Features
+
+* Add optional path var for instance level bucket path ([#251](https://github.com/wandb/terraform-aws-wandb/issues/251)) ([9ccc357](https://github.com/wandb/terraform-aws-wandb/commit/9ccc357c5f991683394efdaa7310d78975ee69b7))
+
+### [4.22.3](https://github.com/wandb/terraform-aws-wandb/compare/v4.22.2...v4.22.3) (2024-08-01)
+
+
+### Bug Fixes
+
+* Bump operator chart versions ([#254](https://github.com/wandb/terraform-aws-wandb/issues/254)) ([9932e33](https://github.com/wandb/terraform-aws-wandb/commit/9932e331a58ee0be67e90f7ffb6391f28c17eac1))
+
+### [4.22.2](https://github.com/wandb/terraform-aws-wandb/compare/v4.22.1...v4.22.2) (2024-08-01)
+
+
+### Bug Fixes
+
+* Test TF Pipeline ([#253](https://github.com/wandb/terraform-aws-wandb/issues/253)) ([b46d6d2](https://github.com/wandb/terraform-aws-wandb/commit/b46d6d2fbac25af8374bb3121b410f2d78935bfc))
+
+### [4.22.1](https://github.com/wandb/terraform-aws-wandb/compare/v4.22.0...v4.22.1) (2024-08-01)
+
+
+### Bug Fixes
+
+* Bump operator chart versions ([#252](https://github.com/wandb/terraform-aws-wandb/issues/252)) ([e8987f2](https://github.com/wandb/terraform-aws-wandb/commit/e8987f29fdde5f9164a1c0133ea2d6e672a1fa5d))
+
+## [4.22.0](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.6...v4.22.0) (2024-07-31)
+
+
+### Features
+
+* Bump operator image and chart versions ([#250](https://github.com/wandb/terraform-aws-wandb/issues/250)) ([1c60818](https://github.com/wandb/terraform-aws-wandb/commit/1c608185dc6dd68d560d7715060a65fc8719c895))
+
+### [4.21.6](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.5...v4.21.6) (2024-07-24)
+
+
+### Bug Fixes
+
+* Always let the node role have access to the `default_kms_key` ([#249](https://github.com/wandb/terraform-aws-wandb/issues/249)) ([d8fa06f](https://github.com/wandb/terraform-aws-wandb/commit/d8fa06f89da48443cb9fe0a45f491e5c13bb41cc))
+
+### [4.21.5](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.4...v4.21.5) (2024-07-24)
+
+
+### Bug Fixes
+
+* Use bucket KMS key arn if provided for W&B managed bucket, always use that key even if empty for customer provided buckets ([#248](https://github.com/wandb/terraform-aws-wandb/issues/248)) ([48131b7](https://github.com/wandb/terraform-aws-wandb/commit/48131b79219071b0a1311bbb5bc468a62c51e266))
+
+### [4.21.4](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.3...v4.21.4) (2024-07-23)
+
+
+### Bug Fixes
+
+* Yace Yace Yace ([#247](https://github.com/wandb/terraform-aws-wandb/issues/247)) ([cf419bd](https://github.com/wandb/terraform-aws-wandb/commit/cf419bdd8d1a3c3996738bbfe8b292579db59d2f))
+
+### [4.21.3](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.2...v4.21.3) (2024-07-23)
+
+
+### Bug Fixes
+
+* YACE scoping ([#246](https://github.com/wandb/terraform-aws-wandb/issues/246)) ([47871c8](https://github.com/wandb/terraform-aws-wandb/commit/47871c846c13ed93ffa71b68c8177f0d2d99d7cf))
+
+### [4.21.2](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.1...v4.21.2) (2024-07-18)
+
+
+### Bug Fixes
+
+* Condition to create kms.aws_kms_grant.clickhouse was incorrect ([#245](https://github.com/wandb/terraform-aws-wandb/issues/245)) ([78d9be7](https://github.com/wandb/terraform-aws-wandb/commit/78d9be7c0b1126aada5e5df7539ae47ecc6b3368))
+
+### [4.21.1](https://github.com/wandb/terraform-aws-wandb/compare/v4.21.0...v4.21.1) (2024-07-18)
+
+
+### Bug Fixes
+
+* Don't create KMS key and related resources for CH by default ([#244](https://github.com/wandb/terraform-aws-wandb/issues/244)) ([42d64ba](https://github.com/wandb/terraform-aws-wandb/commit/42d64bae1847a6d26b16bbf46cd341a39389ad0f))
+
+## [4.21.0](https://github.com/wandb/terraform-aws-wandb/compare/v4.20.3...v4.21.0) (2024-07-17)
+
+
+### Features
+
+* Set up KMS key for clickhouse CMEK and endpoint for PL ([#243](https://github.com/wandb/terraform-aws-wandb/issues/243)) ([1d2fb92](https://github.com/wandb/terraform-aws-wandb/commit/1d2fb921792019b6356e0f89b7c117dda168339a))
+
 ### [4.20.3](https://github.com/wandb/terraform-aws-wandb/compare/v4.20.2...v4.20.3) (2024-07-11)